requirements.txt
Django==4.2.7
djangorestframework==3.14.0
hedera-sdk-py>=2.50.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
requests==2.31.0
Pillow==10.1.0